refactor: split wasm modules
